Update bandTimeVis.m
Works slightly faster because the query is optimized. Filters the data according to the spectral band specified by the user. Takes the average if more than 1 band falls into the selected band. E.g. user wants 760 nm signal, but sensor has 4 bands in between 170 and 171. Then the result will be the average of the four bands.