Create parforMain.m
Connect multiple clients to the DB.
This method first connects to the DB and gets all the hierarchies that are not yet processed (ideally this should come down to one per day).
It then creates as many connections to the DB as possible (# cpu cores). Each worker then processes a hierarchy (i.e. a single day).