Refactoring
With the new direct access to the DB some things needed to be changed.
Mainly the use of the parent_id is introduced. This removes the need to have the user_data.campaign.
However it might still be needed later on when multiple campaigns are present. Because each campaign starts hierarchies at 1.