Integrating recent updates to specchio that improve the speed
Processing L0-->L1 makes use of updated specchio functionality. Some methods now allow to pass basically all ids from a single space (copySpectra and updateSpectrumVectors). However there is room for improvements because the Metadata is still going through a single-attribute single-spectrum loop. This loop currently requires the most time in the processing chain.
L0 to L1 now returns the new measurement vectors, the new ids and the filenames. This can be passed on to L1 --> L2.